Add platform to app store app, rearrange lots of code, renames, need diff compoonents, move all empty states up one layer
