+++ frontend/pages/SoftwarePage/SoftwareAddPage/SoftwareAppStoreVpp/SoftwareVppForm/_styles.scss
@@ -31,32 +31,6 @@
     gap: $pad-medium;
   }
 
-  // enable and no software message styles
-  &__enable-vpp-card,
-  &__no-software-card {
-    padding: $pad-xxxlarge auto;
-  }
-
-  &__enable-vpp-message,
-  &__no-software-message {
-    display: flex;
-    flex-direction: column;
-    gap: $pad-small;
-    align-items: center;
-  }
-
-  &__enable-vpp-title,
-  &__no-software-title {
-    font-weight: $bold;
-    margin: 0;
-  }
-
-  &__enable-vpp-description,
-  &__no-software-description {
-    margin: 0;
-    color: $ui-fleet-black-75;
-  }
-
   // list styles
   &__list-container {
     border: 1px solid $ui-fleet-black-10;
