+++ frontend/pages/SoftwarePage/SoftwareAddPage/SoftwareAppStoreVpp/_styles.scss
@@ -1,6 +1,6 @@
 .software-app-store-vpp {
-
-  &__premium-message, &__error {
+  &__premium-message,
+  &__error {
     margin-top: $pad-xxxlarge;
   }
 
@@ -8,5 +8,31 @@
     margin-top: $pad-xxlarge;
     display: flex;
     flex-direction: column;
+  }
+
+  // enable and no vpp message styles
+  &__enable-vpp-card,
+  &__no-vpp-card {
+    padding: $pad-xxxlarge auto;
+  }
+
+  &__enable-vpp-message,
+  &__no-vpp-message {
+    display: flex;
+    flex-direction: column;
+    gap: $pad-small;
+    align-items: center;
+  }
+
+  &__enable-vpp-title,
+  &__no-vpp-title {
+    font-weight: $bold;
+    margin: 0;
+  }
+
+  &__enable-vpp-description,
+  &__no-vpp-description {
+    margin: 0;
+    color: $ui-fleet-black-75;
   }
 }
