+++ frontend/pages/SoftwarePage/SoftwareAddPage/SoftwareFleetMaintained/FleetMaintainedAppDetailsPage/FleetAppDetailsForm/FleetAppDetailsForm.tsx
@@ -6,23 +6,20 @@ import Checkbox from "components/forms/fields/Checkbox";
 import TooltipWrapper from "components/TooltipWrapper";
 import RevealButton from "components/buttons/RevealButton";
 import Button from "components/buttons/Button";
-import Radio from "components/forms/fields/Radio";
 import TargetLabelSelector from "components/TargetLabelSelector";
-import InfoBanner from "components/InfoBanner";
-import CustomLink from "components/CustomLink";
-import { LEARN_MORE_ABOUT_BASE_LINK } from "utilities/constants";
+import {
+  CUSTOM_TARGET_OPTIONS,
+  generateHelpText,
+  InstallType,
+  InstallTypeSection,
+} from "pages/SoftwarePage/helpers";
 
 import AdvancedOptionsFields from "pages/SoftwarePage/components/AdvancedOptionsFields";
 
-import {
-  CUSTOM_TARGET_OPTIONS,
-  generateFormValidation,
-  generateHelpText,
-} from "./helpers";
+import { generateFormValidation } from "./helpers";
 
 const baseClass = "fleet-app-details-form";
 
-export type InstallType = "manual" | "automatic";
 export interface IFleetMaintainedAppFormData {
   selfService: boolean;
   installScript: string;
@@ -52,95 +49,6 @@ interface IFleetAppDetailsFormProps {
   onCancel: () => void;
   onSubmit: (formData: IFleetMaintainedAppFormData) => void;
 }
-
-interface IInstallTypeSection {
-  installType: InstallType;
-  onChangeInstallType: (value: string) => void;
-  isCustomPackage?: boolean;
-  isExeCustomPackage?: boolean;
-}
-
-// Also used in custom package form (PackageForm.tsx)
-export const InstallTypeSection = ({
-  installType,
-  onChangeInstallType,
-  isCustomPackage = false,
-  isExeCustomPackage = false,
-}: IInstallTypeSection) => {
-  const isAutomaticDisabled = isExeCustomPackage;
-  const AUTOMATIC_DISABLED_TOOLTIP = (
-    <>
-      Fleet can&apos;t create a policy to detect existing installations for
-      <br /> .exe packages. To automatically install an .exe, add a custom
-      <br /> policy and enable the install software automation on the
-      <br /> <b>Policies</b> page.
-    </>
-  );
-
-  return (
-    <fieldset>
-      <legend>Install</legend>
-      <div className={`${baseClass}__radio-input`}>
-        <Radio
-          checked={installType === "manual"}
-          id="manual-radio-btn"
-          value="manual"
-          name="install-type"
-          label="Manual"
-          onChange={onChangeInstallType}
-          helpText={
-            <>
-              Manually install on the <b>Host details</b> page for each host.
-            </>
-          }
-        />
-        <Radio
-          checked={installType === "automatic"}
-          id="automatic-radio-btn"
-          value="automatic"
-          name="install-type"
-          label="Automatic"
-          disabled={isAutomaticDisabled}
-          tooltip={isAutomaticDisabled && AUTOMATIC_DISABLED_TOOLTIP}
-          onChange={onChangeInstallType}
-          helpText={
-            <>
-              Automatically install on each host that&apos;s{" "}
-              <TooltipWrapper
-                tipContent={
-                  <>
-                    If the host already has any version of this
-                    <br /> software, it won&apos;t be installed.
-                  </>
-                }
-              >
-                missing this software
-              </TooltipWrapper>
-              . Policy that triggers install can be customized after software is
-              added.
-            </>
-          }
-        />
-      </div>
-      {installType === "automatic" && isCustomPackage && (
-        <InfoBanner
-          color="yellow"
-          cta={
-            <CustomLink
-              url={`${LEARN_MORE_ABOUT_BASE_LINK}/query-templates-for-automatic-software-install`}
-              text="Learn more"
-              newTab
-            />
-          }
-        >
-          Installing software over existing installations might cause issues.
-          Fleet&apos;s policy may not detect these existing installations.
-          Please create a test team in Fleet to verify a smooth installation.
-        </InfoBanner>
-      )}
-    </fieldset>
-  );
-};
 
 const FleetAppDetailsForm = ({
   labels,
@@ -237,6 +145,7 @@ const FleetAppDetailsForm = ({
   return (
     <form className={baseClass} onSubmit={onSubmitForm}>
       <InstallTypeSection
+        className={baseClass}
         installType={formData.installType}
         onChangeInstallType={onChangeInstallType}
       />
