+++ frontend/pages/SoftwarePage/SoftwareAddPage/SoftwareFleetMaintained/FleetMaintainedAppDetailsPage/FleetAppDetailsForm/helpers.tsx
@@ -1,7 +1,3 @@
-import React from "react";
-
-import { IDropdownOption } from "interfaces/dropdownOption";
-
 // @ts-ignore
 import validateQuery from "components/forms/validators/validate_query";
 
@@ -94,46 +90,4 @@ export const generateFormValidation = (
   });
 
   return formValidation;
-};
-
-export const CUSTOM_TARGET_OPTIONS: IDropdownOption[] = [
-  {
-    value: "labelsIncludeAny",
-    label: "Include any",
-    disabled: false,
-  },
-  {
-    value: "labelsExcludeAny",
-    label: "Exclude any",
-    disabled: false,
-  },
-];
-
-export const generateHelpText = (installType: string, customTarget: string) => {
-  if (customTarget === "labelsIncludeAny") {
-    return installType === "manual" ? (
-      <>
-        Software will only be available for install on hosts that{" "}
-        <b>have any</b> of these labels:
-      </>
-    ) : (
-      <>
-        Software will only be installed on hosts that <b>have any</b> of these
-        labels:
-      </>
-    );
-  }
-
-  // this is the case for labelsExcludeAny
-  return installType === "manual" ? (
-    <>
-      Software will only be available for install on hosts that{" "}
-      <b>don&apos;t have any</b> of these labels:
-    </>
-  ) : (
-    <>
-      Software will only be installed on hosts that <b>don&apos;t have any</b>{" "}
-      of these labels:{" "}
-    </>
-  );
 };
