+++ frontend/pages/SoftwarePage/components/PackageForm/helpers.tsx
@@ -1,9 +1,3 @@
-import React from "react";
-
-import { IDropdownOption } from "interfaces/dropdownOption";
-import { IAppStoreApp, ISoftwarePackage } from "interfaces/software";
-import { InstallType } from "pages/SoftwarePage/SoftwareAddPage/SoftwareFleetMaintained/FleetMaintainedAppDetailsPage/FleetAppDetailsForm/FleetAppDetailsForm";
-
 // @ts-ignore
 import validateQuery from "components/forms/validators/validate_query";
 
@@ -107,98 +101,3 @@ export const generateFormValidation = (formData: IPackageFormData) => {
 };
 
 export default generateFormValidation;
-
-export const CUSTOM_TARGET_OPTIONS: IDropdownOption[] = [
-  {
-    value: "labelsIncludeAny",
-    label: "Include any",
-    disabled: false,
-  },
-  {
-    value: "labelsExcludeAny",
-    label: "Exclude any",
-    disabled: false,
-  },
-];
-
-export const getInstallType = (
-  softwarePackage: ISoftwarePackage
-): InstallType => {
-  return softwarePackage.automatic_install_policies ? "automatic" : "manual";
-};
-
-export const getTargetType = (
-  softwareInstaller: ISoftwarePackage | IAppStoreApp
-) => {
-  if (!softwareInstaller) return "All hosts";
-
-  return !softwareInstaller.labels_include_any &&
-    !softwareInstaller.labels_exclude_any
-    ? "All hosts"
-    : "Custom";
-};
-
-export const getCustomTarget = (
-  softwareInstaller: ISoftwarePackage | IAppStoreApp
-) => {
-  if (!softwareInstaller) return "labelsIncludeAny";
-
-  return softwareInstaller.labels_include_any
-    ? "labelsIncludeAny"
-    : "labelsExcludeAny";
-};
-
-export const generateSelectedLabels = (
-  softwareInstaller: ISoftwarePackage | IAppStoreApp
-) => {
-  if (
-    !softwareInstaller ||
-    (!softwareInstaller.labels_include_any &&
-      !softwareInstaller.labels_exclude_any)
-  ) {
-    return {};
-  }
-
-  const customTypeKey = softwareInstaller.labels_include_any
-    ? "labels_include_any"
-    : "labels_exclude_any";
-
-  return (
-    softwareInstaller[customTypeKey]?.reduce<Record<string, boolean>>(
-      (acc, label) => {
-        acc[label.name] = true;
-        return acc;
-      },
-      {}
-    ) ?? {}
-  );
-};
-
-export const generateHelpText = (installType: string, customTarget: string) => {
-  if (customTarget === "labelsIncludeAny") {
-    return installType === "manual" ? (
-      <>
-        Software will only be available for install on hosts that{" "}
-        <b>have any</b> of these labels:
-      </>
-    ) : (
-      <>
-        Software will only be installed on hosts that <b>have any</b> of these
-        labels:
-      </>
-    );
-  }
-
-  // this is the case for labelsExcludeAny
-  return installType === "manual" ? (
-    <>
-      Software will only be available for install on hosts that{" "}
-      <b>don&apos;t have any</b> of these labels:
-    </>
-  ) : (
-    <>
-      Software will only be installed on hosts that <b>don&apos;t have any</b>{" "}
-      of these labels:{" "}
-    </>
-  );
-};
